% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{add_user}
\alias{add_user}
\title{Adds an user to the database}
\usage{
add_user(con, salt, username, password)
}
\arguments{
\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}

\item{username}{the username}

\item{password}{the password}
}
\value{
the result of the query or -1 if user already exists
}
\description{
Adds an user to the database
}
