% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chgpwd.R
\name{chgpwd}
\alias{chgpwd}
\title{Change password module, Server part}
\usage{
chgpwd(input, output, session, con, salt, username)
}
\arguments{
\item{input}{for the \code{callmodule}}

\item{output}{for the \code{callmodule}}

\item{session}{for the \code{callmodule}}

\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}

\item{username}{the username}
}
\description{
Change password module, Server part
}
