% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{create_users_table}
\alias{create_users_table}
\title{Creates the user table}
\usage{
create_users_table(con, salt)
}
\arguments{
\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}
}
\value{
the result of the transaction
}
\description{
A Helper function to create the user table in the database
It adds also the admin and guest users with admin and guest passwords
respectively.
}
