% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{delete_user}
\alias{delete_user}
\title{Deletes an user from the database}
\usage{
delete_user(con, username)
}
\arguments{
\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{username}{the username}
}
\value{
the result of the executed sentence or -1 if the user does not exists
}
\description{
Deletes an user from the database
}
