% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{get_con}
\alias{get_con}
\title{Get a connection string}
\usage{
get_con(configname = "defaultdb")
}
\arguments{
\item{configname}{section of the \code{config.yml} file with the parameters}
}
\value{
a \code{\link[pool]{dbPool}} connection object
}
\description{
Read a \code{config.yml} file to obtain the
connection parameter string to create a \code{\link[pool]{dbPool}}
connection object
}
