% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{is_admin}
\alias{is_admin}
\title{Test if the user is admin}
\usage{
is_admin(username)
}
\arguments{
\item{username}{the username}
}
\value{
TRUE if the user is admin
}
\description{
Test if the user is admin
}
