% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{is_authorized}
\alias{is_authorized}
\title{Is the user authorized?}
\usage{
is_authorized(con, salt, username, password)
}
\arguments{
\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}

\item{username}{the username}

\item{password}{the password}
}
\value{
TRUE if the user is authorized, FALSE otherwise
}
\description{
Return true if the user/password combination is found in the database
}
