% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{is_guest}
\alias{is_guest}
\title{Test if the user is guest}
\usage{
is_guest(username)
}
\arguments{
\item{username}{the username}
}
\value{
TRUE if the user is guest
}
\description{
Test if the user is guest
}
