% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management.R
\name{modify_password}
\alias{modify_password}
\title{Modifies the password for an user}
\usage{
modify_password(con, salt, username, password)
}
\arguments{
\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}

\item{username}{the username}

\item{password}{the password}
}
\value{
the result of the executed sentence or -1 if the user does not exists
}
\description{
Modifies the password for an user
}
