% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passprotServer.R
\name{passprotServer}
\alias{passprotServer}
\title{Return a function that can be use as Server for Shiny app}
\usage{
passprotServer(con, salt, appmodule, ...)
}
\arguments{
\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}

\item{appmodule}{a \code{Shiny module server} object}

\item{...}{extra parameters for \code{appmodule}}
}
\value{
a function with \code{input}, \code{output}, \code{session} arguments
}
\description{
This wrap have the logic for the user management and authorization
The logic of the user application should be defined in the \code{appmodule}
}
