% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passprotUI.R
\name{passprotUI}
\alias{passprotUI}
\title{Creates the UI}
\usage{
passprotUI(apptitle, apptheme = "cerulean", appmoduleUI, ...)
}
\arguments{
\item{apptitle}{application title}

\item{apptheme}{\code{Boostrap} theme from \code{shinythemes}}

\item{appmoduleUI}{a module UI object}

\item{...}{Other parameters for the \code{appModuleUI}}
}
\value{
a \code{\link[shiny]{fluidPage}} object
}
\description{
Returns a UI page with the user management UI. The user application
UI should be defined in the \code{appmoduleUI}
}
