% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chgpwd.R
\name{chgpwd}
\alias{chgpwd}
\title{Change passwod module, Server part}
\usage{
chgpwd(input, output, session, con, salt, username)
}
\arguments{
\item{input}{for the callmodule}

\item{output}{for the callmodule}

\item{session}{for the callmodule}

\item{con}{\code{\link[pool]{dbPool}} connection object}

\item{salt}{a string to improve masking of passwords}

\item{username}{the username}
}
\description{
Change passwod module, Server part
}
